parser/tests/tree_construction/basic.dat

#data
<!doctype html><p>foo<address>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “address”.
#document
| <!DOCTYPE html>
|  <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|     <address>
|       "bar"
|       <p>
|         "baz"
